+++ Package.swift
@@ -4,40 +4,31 @@
 import PackageDescription
 
 let package = Package(
-    name: "flexiBLE_signal",
+    name: "FlexiBLESignal",
     platforms: [
         .macOS(.v12),
-        .iOS(.v13),
+        .iOS(.v15),
         .tvOS(.v13),
         .watchOS(.v6)
     ],
     products: [
         // Products define the executables and libraries a package produces, and make them visible to other packages.
         .library(
-            name: "flexiBLE_signal",
-            targets: ["flexiBLE_signal"])
+            name: "FlexiBLESignal",
+            targets: ["FlexiBLESignal"]),
     ],
     dependencies: [
         // Dependencies declare other packages that this package depends on.
         // .package(url: /* package url */, from: "1.0.0"),
-        .package(
-              url: "https://github.com/apple/swift-collections.git",
-              .upToNextMajor(from: "1.0.3")
-            )
     ],
     targets: [
         // Targets are the basic building blocks of a package. A target can define a module or a test suite.
         // Targets can depend on other targets in this package, and on products in packages this package depends on.
         .target(
-            name: "flexiBLE_signal",
-            dependencies: [
-                .product(name: "Collections", package: "swift-collections")
-            ]),
+            name: "FlexiBLESignal",
+            dependencies: []),
         .testTarget(
-            name: "flexiBLE_signalTests",
-            dependencies: [
-                "flexiBLE_signal",
-                .product(name: "Collections", package: "swift-collections")
-            ]),
+            name: "FlexiBLESignalTests",
+            dependencies: ["FlexiBLESignal"]),
     ]
 )
