+++ README.md
@@ -1,3 +1,3 @@
-# flexiBLE-signal
+# FlexiBLESignal
 
 A description of this package.
