+++ Sources/FlexiBLESignal/FlexiBLESignal.swift
@@ -1,0 +1,12 @@
+import Foundation
+import Accelerate
+
+//public struct Point<T: FloatingPoint, U: FloatingPoint> {
+//    public var x: T
+//    public var y: U
+//}
+//
+//public enum Precision {
+//    case single
+//    case double
+//}
