--- Tests/flexiBLE_signalTests/flexiBLE_signalTests.swift
@@ -1,11 +1,0 @@
-import XCTest
-@testable import flexiBLE_signal
-
-final class flexiBLE_signalTests: XCTestCase {
-    func testExample() throws {
-        // This is an example of a functional test case.
-        // Use XCTAssert and related functions to verify your tests produce the correct
-        // results.
-        XCTAssertEqual(flexiBLE_signal().text, "Hello, World!")
-    }
-}
