--- playgrounds/signal_gen.playground/Contents.swift
@@ -1,4 +1,0 @@
-import Foundation
-import Accelerate
-
-SinWave
